refactor(playback): extract PlaybackRepository coordinators under ≤1000 LOC
Move chapters/transcript, queue, media bridge, history, and transport logic
into dedicated playback package coordinators while keeping PlaybackRepository
as a public façade. Add PlaybackSessionMapping helper + unit tests.
